+++ lib/screens/app/ecosystem/invites/create_invite.dart
@@ -5,7 +5,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:provider/provider.dart';
 import 'package:seeds/constants/app_colors.dart';
 import 'package:seeds/i18n/invites.i18n.dart';
-import 'package:seeds/providers/services/eos_service.dart';
+import 'package:seeds/providers/services/eos/eos_service.dart';
 import 'package:seeds/providers/services/links_service.dart';
 import 'package:seeds/utils/invites.dart';
 import 'package:seeds/widgets/available_balance.dart';
