+++ test/unit_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:seeds/constants/http_mock_response.dart';
-import 'package:seeds/providers/services/eos_service.dart';
+import 'package:seeds/providers/services/eos/eos_service.dart';
 import 'package:seeds/providers/services/http_service.dart';
 import 'package:seeds/utils/invites.dart';
